20230329 0330DEMO用的，刪除一些輸出的東西，統一繪製風格 修了一些小邏輯 目前已知問題： 1.Group物件可以連線 2.Group物件內無法連線 3.多層結構下移動由於是遞迴call的，所以下層的結構移動距離會加倍

+++ UML_Editor_Java_2023/src/UML_Object/ShapeClass.java
@@ -28,9 +28,6 @@ public class ShapeClass extends BasicShape{
             g.fillRect(x1, y1, width, height);
             g.setColor(Color.white);
         }
-
-
-
         g.drawRect(x1, y1, width, height);
 
         int portion = height ;
@@ -40,7 +37,7 @@ public class ShapeClass extends BasicShape{
 
         // find the width for the specified string.
         int stringWidth = g.getFontMetrics(font).stringWidth(objName);
-        double empty = (Math.abs(x1-x2) - stringWidth)/2;
+        double empty = (float)(Math.abs(x1-x2) - stringWidth)/2;
         g.setFont(font);
         g.drawString(objName, x1 + (int)empty, y1 + 25);
     }
